refactor(user): moved bpf loading in library
